+++ src/index.ts
@@ -5,3 +5,4 @@ export * from './types';
 export * from './db';
 export * from './models';
 export * from './errors';
+export * from './command';